requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# CUDA 12.1 PyTorch build
torch==2.3.0+cu121
torchvision==0.18.0+cu121

# OCR and image processing
easyocr==1.7.2
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==12.1.1

# Parsing and CLI
google-genai==1.73.1
python-dotenv==1.2.2
tqdm==4.67.3

# Tests
pytest==9.0.3
